+++ src/app/app-routing.module.ts
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2025 gematik GmbH
+    Copyright (c) 2026 gematik GmbH
     Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
     European Commission – subsequent versions of the EUPL (the "Licence").
     You may not use this work except in compliance with the Licence.
@@ -15,15 +15,15 @@
     find details in the "Readme" file.
  */
 
+import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HashLocationStrategy, LocationStrategy } from '@angular/common';
-import { IgsMeldungComponent } from './components/igs-meldung/igs-meldung.component';
+import { AppWrapperComponent } from './app-wrapper.component';
 
 const routes: Routes = [
   {
     path: '**',
-    component: IgsMeldungComponent,
+    component: AppWrapperComponent,
   },
 ];